+++ modules/graphical-minimal.nix
@@ -23,9 +23,8 @@
     '';
 
     # WORKAROUND: Screensaver starts on the login screen and cannot be closed from VNC
-    system.activationScripts.extraActivation.text = ''
-      defaults write /Library/Preferences/com.apple.screensaver loginWindowIdleTime 0
-    '';
+    system.defaults.CustomSystemPreferences."/Library/Preferences/com.apple.screensaver".loginWindowIdleTime =
+      0;
   };
 
   nixosModule = { user, pkgs, ... }: {