+++ secrets/secrets.nix
@@ -10,6 +10,11 @@ in {
     inherit (hosts) phi;
   };
 
+  "github-runner.age".publicKeys = builtins.attrValues {
+    inherit (users) enzime;
+    inherit (hosts) echo;
+  };
+
   "nextcloud.age".publicKeys = builtins.attrValues {
     inherit (users) enzime;
     inherit (hosts) phi;